+++ src/StyledComponents/BackPageButton/BackPageButton.jsx
@@ -1,16 +1,15 @@
 import { Button } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-import { PublicRoutes } from './../../models/routes';
 import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
 
 
-function BackPageButton() {
+function BackPageButton({_urlBase = ""}) {
 
   const navigate = useNavigate();
 
   const goBack = (e) => {
     e.preventDefault();
-    navigate(`/${PublicRoutes.HOME}`);
+    navigate(`${_urlBase}`);
   }
 
   return (
