+++ src/models/Topics.jsx
@@ -1,0 +1,24 @@
+import { TopicsRoutes } from "./routes";
+
+export const AllTopics = [
+  {
+    name: "Varianza",
+    url: TopicsRoutes.VARIANCE
+  },
+  {
+    name: "Media",
+    url: TopicsRoutes.MEDIA
+  },
+  {
+    name: "Mediana",
+    url: TopicsRoutes.MEDIANA
+  },
+  {
+    name: "Moda",
+    url: TopicsRoutes.MODA
+  },
+  {
+    name: "Error",
+    url: TopicsRoutes.ERROR
+  }
+];
