+++ src/models/routes.jsx
@@ -1,10 +1,74 @@
+/*
+
+@Abbreviations
+
+MCT Medidas de Tendencia Central
+
+*/
 export const PublicRoutes = {
+  HOME: "",
+  CALCULATOR: "calculator",
+  SAMPLING: "sampling",
+  PROBABILITIES: "probabilities",
+  DISTRIBUTION_TABLES: "distributions",
+  SEARCH_SYMBOL: "symbols",
+  TOPICS: "topics",
+  SYMBOLS: "symbols"
+}
+
+/*
+
+@Abbreviations
+
+DIST Distribuciones
+
+*/
+
+export const DistTablesRoutes = {
+  HOME: "",
+  DISTZ: "valuez",
+  DISTF: "valuef"
+}
+
+export const CalculatorRoutes = {
   HOME: "",
   FRECUENCY_TABLES: "tables",
   MCT: "mct",
   VARIANCE: "varianceanderror",
-  PERCENTILES: "percentile",
-  SYMBOLS: "symbols",
-  SEARCHZ: "searchz",
-  GRAPHS: "graphs"
+  PERCENTILES: "percentile"
+}
+
+/*
+@Abbreviations
+
+MAS Muestreo Aleatorio Simple
+MAE Muestreo Aleatorio Estratificado
+CS Cluster Sampling / Muestreo por conglomerados
+SS Systematic Sampling / Muestreo Sistemático
+MD Double Sampling / Muestreo Doble
+
+*/
+
+export const SamplingRoutes = {
+  HOME: "",
+  MAS: "mas",
+  MAE: "mae",
+  MPC: "mpc",
+  MS: "ms",
+  MD: "md"
+}
+
+export const ProbabilitiesRoutes = {
+  HOME: "",
+  PERMUTATIONS: "permutations",
+  COMBINATIONS: "combinations"
+}
+
+export const TopicsRoutes = {
+  HOME: "",
+  VARIANCE: "variance",
+  ERROR: "error",
+  MEDIA: "media",
+  MEDIANA: "mediana",
+  MODA: "moda",
 }
