+++ src/pages/Calculator/MCT/Components/MCT_description.jsx
@@ -1,0 +1,13 @@
+import { Box, Typography } from "@mui/material";
+import { MCTDescription } from "../../../../Descriptions/DescriptionsAllTopics/descriptions";
+
+function MCT_description() {
+  return (
+    <Box component="div" className="p-3">
+      <Typography variant="inherit">
+        {MCTDescription.fullInfo}
+      </Typography>
+    </Box>
+  );
+}
+export default MCT_description;
