+++ src/pages/MainHome/Components/Bell.jsx
@@ -1,0 +1,7 @@
+
+function Bell() {
+  return (
+    <div>Bell</div>
+  );
+}
+export default Bell;
