+++ src/pages/StatGraphs/StatGraphs.jsx
@@ -17,7 +17,6 @@ import { Bar } from 'react-chartjs-2';
 import { Box, Typography, useTheme } from '@mui/material';
 
 import RechartsTest from './Components/RechartsTest';
-import BackPageButton from '../../StyledComponents/BackPageButton/BackPageButton';
 
 function StatGraphs(props) {
   const theme = useTheme();
@@ -89,9 +88,6 @@ function StatGraphs(props) {
 
   return (
     <Box component="div" className="container-lg">
-      <Box component="div" className="p-3">
-        <BackPageButton />
-      </Box>
       <Typography variant="h5" className="d-flex align-items-center justify-content-center p-4">Gráficos Estadísticos</Typography>
       <Box
         sx={{
