+++ src/pages/TopicTeory/Error/ErrorTeory.jsx
@@ -1,0 +1,7 @@
+
+function ErrorTeory() {
+  return (
+    <div>ErrorTeory</div>
+  )
+}
+export default ErrorTeory;
