+++ src/pages/TopicTeory/Media/MediaTeory.jsx
@@ -1,0 +1,7 @@
+
+function MediaTeory() {
+  return (
+    <div>MediaTeory</div>
+  )
+}
+export default MediaTeory;
