+++ src/pages/TopicTeory/Mediana/MedianaTeory.jsx
@@ -1,0 +1,7 @@
+
+function MedianaTeory() {
+  return (
+    <div>MedianaTeory</div>
+  )
+}
+export default MedianaTeory;
