+++ src/pages/TopicTeory/Moda/ModaTeory.jsx
@@ -1,0 +1,7 @@
+
+function ModaTeory() {
+  return (
+    <div>ModaTeory</div>
+  )
+}
+export default ModaTeory;
