+++ src/pages/TopicTeory/Variance/VarianceTeory.jsx
@@ -1,0 +1,7 @@
+
+function VarianceTeory() {
+  return (
+    <div>VarianceTeory</div>
+  )
+}
+export default VarianceTeory;
